requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Document processing (required by code)
PyMuPDF==1.23.8
opencv-python-headless==4.8.1.78
python-docx==1.1.0
extract-msg==0.47.0
beautifulsoup4==4.12.2
pillow==10.1.0
pdfkit==1.0.0

# ML/AI - CPU versions with compatible deps  
# torch installed separately via: pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
sentence-transformers==2.2.2
transformers==4.21.1
huggingface-hub==0.10.1
pinecone==7.3.0
google-generativeai==0.3.2

# Data processing
scikit-learn==1.3.2
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1